+++ R/preprocess-sanitize.R
@@ -15,7 +15,13 @@ sanitize_plan <- function(plan, allow_duplicated_targets = FALSE) {
   if (!allow_duplicated_targets) {
     plan <- assert_unique_targets(plan[, cols])
   }
-  arrange_plan_cols(plan)
+  plan <- arrange_plan_cols(plan)
+  for (col in lang_cols(plan)) {
+    if (!is.list(plan[[col]])) {
+      plan[[col]] <- lapply(plan[[col]], safe_parse)
+    }
+  }
+  as_drake_plan(plan)
 }
 
 sanitize_targets <- function(plan, targets) {