+++ tests/testthat/test-map-reduce.R
@@ -14,10 +14,10 @@ test_with_dir("map_plan()", {
   plan3 <- map_plan(args = args, fun = fn, id = x)
   plan4 <- map_plan(args = args, fun = "fn", id = "x", character_only = TRUE)
   plan5 <- map_plan(args = args, fun = fn, id = x, trace = TRUE)
-  expect_equal(plan1$command, plan2$command)
-  expect_equal(plan2, plan3)
-  expect_equal(plan3, plan4)
-  expect_equal(weak_as_tibble(cbind(plan3, args)), plan5)
+  expect_equal(deparse_lang_col(plan1$command), deparse_lang_col(plan2$command))
+  equivalent_plans(plan2, plan3)
+  equivalent_plans(plan3, plan4)
+  equivalent_plans(weak_as_tibble(cbind(plan3, args)), plan5)
   cache <- storr::storr_environment()
   make(plan2, session_info = FALSE, cache = cache)
   expect_equal(
@@ -67,18 +67,20 @@ test_with_dir("map_plan() with symbols", {
 test_with_dir("gather_plan()", {
   df <- drake_plan(data = simulate(center = MU, scale = SIGMA))
   m0 <- evaluate_plan(df, wildcard = "NULL", values = 1:2)
-  expect_equal(m0, df)
+  equivalent_plans(m0, df)
   m1 <- evaluate_plan(df, rules = list(nothing = 1:2), expand = FALSE)
-  expect_equal(m1, df)
+  equivalent_plans(m1, df)
   x <- expand_plan(df, values = c("rep1", "rep2"))
   x6 <- gather_plan(x, append = FALSE)
-  y <- weak_tibble(
-    target = "target",
-    command = "list(data_rep1 = data_rep1, data_rep2 = data_rep2)"
+  y <- sanitize_plan(
+    weak_tibble(
+      target = "target",
+      command = "list(data_rep1 = data_rep1, data_rep2 = data_rep2)"
+    )
   )
-  expect_equal(x6, y)
+  equivalent_plans(x6, y)
   z <- gather_plan(x, append = TRUE)
-  expect_equal(z, bind_plans(x, y))
+  equivalent_plans(z, bind_plans(x, y))
   x7 <- gather_plan(
     x, target = "my_summaries", gather = "rbind", append = FALSE
   )
@@ -86,7 +88,7 @@ test_with_dir("gather_plan()", {
     target = "my_summaries",
     command = "rbind(data_rep1 = data_rep1, data_rep2 = data_rep2)"
   )
-  expect_equal(x7, y)
+  equivalent_plans(x7, y)
 })
 
 test_with_dir("reduce_plan()", {
@@ -105,13 +107,13 @@ test_with_dir("reduce_plan()", {
     target = "x_sum",
     command = paste0(x_plan$target, collapse = " + ")
   )
-  expect_equal(x, x0)
+  equivalent_plans(x, x0)
   z <- reduce_plan(
     x_plan, target = "x_sum", pairwise = FALSE,
     begin = "", end = "", append = TRUE
   )
   z0 <- bind_plans(x_plan, x)
-  expect_equal(z, z0)
+  equivalent_plans(z, z0)
   make(rbind(x_plan, x), session_info = FALSE)
   expect_equal(readd(x_sum), sum(1:8))
   clean(destroy = TRUE)
@@ -126,7 +128,7 @@ test_with_dir("reduce_plan()", {
       "x_sum_5 + x_sum_6"
     )
   )
-  expect_equal(x, x0)
+  equivalent_plans(x, x0)
   x <- reduce_plan(
     x_plan, target = "x_sum", pairwise = FALSE,
     begin = "", end = ""
@@ -135,7 +137,7 @@ test_with_dir("reduce_plan()", {
     target = "x_sum",
     command = paste0(x_plan$target, collapse = " + ")
   )
-  expect_equal(x, x0)
+  equivalent_plans(x, x0)
   x <- reduce_plan(x_plan, target = "x_sum", pairwise = TRUE)
   x0 <- weak_tibble(
     target = c(paste0("x_sum_", 1:6), "x_sum"),
@@ -145,7 +147,7 @@ test_with_dir("reduce_plan()", {
       "x_sum_5 + x_sum_6"
     )
   )
-  expect_equal(x, x0)
+  equivalent_plans(x, x0)
   make(rbind(x_plan, x), session_info = FALSE)
   expect_equal(readd(x_sum), sum(1:8))
   clean(destroy = TRUE)
@@ -166,7 +168,7 @@ test_with_dir("reduce_plan()", {
       "x_sum_6 + x_sum_7"
     )
   )
-  expect_equal(x, x0)
+  equivalent_plans(x, x0)
   make(rbind(x_plan, x), session_info = FALSE)
   expect_equal(readd(x_sum), sum(1:9))
   clean(destroy = TRUE)
@@ -190,7 +192,7 @@ test_with_dir("reduce_plan()", {
       "fun(x_sum_5, x_sum_6)"
     )
   )
-  expect_equal(x, x0)
+  equivalent_plans(x, x0)
   make(rbind(x_plan, x))
   out <- fun(
     fun(
@@ -216,18 +218,15 @@ test_with_dir("gather_by()", {
     trace = TRUE
   )
   x <- gather_by(plan, append = TRUE)
-  y <- weak_tibble(
-    target = c(plan$target, "target"),
-    command = c(
-      plan$command,
-      "list(x_1 = x_1, x_2 = x_2, y_a = y_a, y_b = y_b, z = z)"
-    )
+  new_row <- drake_plan(
+    target = list(x_1 = x_1, x_2 = x_2, y_a = y_a, y_b = y_b, z = z)
   )
-  expect_equivalent(x[, c("target", "command")], y)
+  y <- bind_plans(plan, new_row)
+  equivalent_plans(x[, c("target", "command")], y)
   x <- gather_by(plan, append = TRUE, sep = ".")
-  expect_equivalent(x[, c("target", "command")], y)
+  equivalent_plans(x[, c("target", "command")], y)
   z <- gather_by(plan, append = FALSE)
-  expect_equivalent(z[, c("target", "command")], y[nrow(y), ])
+  equivalent_plans(z[, c("target", "command")], y[nrow(y), ])
   x <- gather_by(
     plan,
     n___from,
@@ -244,7 +243,8 @@ test_with_dir("gather_by()", {
     n__ = NA,
     n___from = c("y", NA)
   )
-  expect_equivalent(x, bind_plans(plan, y))
+  y <- sanitize_plan(y)
+  equivalent_plans(x, bind_plans(plan, y))
   x <- gather_by(plan, n___from, prefix = "xyz", gather = "c", append = TRUE)
   y <- weak_tibble(
     target = c("xyz_y", "xyz_NA"),
@@ -254,8 +254,9 @@ test_with_dir("gather_by()", {
     n__ = NA,
     n___from = c("y", NA)
   )
+  y <- sanitize_plan(y)
   expected <- bind_plans(plan, y)
-  expect_equivalent(x[order(x$target), ], expected[order(expected$target), ])
+  equivalent_plans(x[order(x$target), ], expected[order(expected$target), ])
   x <- gather_by(plan, m__, n__, prefix = "xyz", gather = "c", append = TRUE)
   y <- weak_tibble(
     target = c("xyz_1_NA", "xyz_2_NA", "xyz_NA_a", "xyz_NA_b", "xyz_NA_NA"),
@@ -271,8 +272,9 @@ test_with_dir("gather_by()", {
     n__ = c(NA, NA, "a", "b", NA),
     n___from = as.character(NA)
   )
+  y <- sanitize_plan(y)
   expected <- bind_plans(plan, y)
-  expect_equivalent(x[order(x$target), ], expected[order(expected$target), ])
+  equivalent_plans(x[order(x$target), ], expected[order(expected$target), ])
   plan$n___from <- c("x", "x", "y", "y", NA)
   x <- gather_by(
     plan,
@@ -282,11 +284,9 @@ test_with_dir("gather_by()", {
     append = TRUE,
     filter = n___from == "x"
   )
-  y <- weak_tibble(
-    target = c(plan$target, "xyz_x"),
-    command = c(plan$command, "c(x_1 = x_1, x_2 = x_2)")
-  )
-  expect_equivalent(x[, c("target", "command")], y)
+  new_row <- drake_plan(xyz_x = c(x_1 = x_1, x_2 = x_2))
+  y <- bind_plans(plan, new_row)
+  equivalent_plans(x[, c("target", "command")], y)
 })
 
 test_with_dir("reduce_by()", {
@@ -300,13 +300,11 @@ test_with_dir("reduce_by()", {
     trace = TRUE
   )
   x <- reduce_by(plan, pairwise = FALSE, append = TRUE)
-  y <- weak_tibble(
-    target = c(plan$target, "target"),
-    command = c(plan$command, "x_1 + x_2 + x_3 + x_4 + y_a + y_b + z")
-  )
-  expect_equivalent(x[, c("target", "command")], y)
+  new_row <- drake_plan(target = x_1 + x_2 + x_3 + x_4 + y_a + y_b + z)
+  y <- bind_plans(plan, new_row)
+  equivalent_plans(x[, c("target", "command")], y)
   z <- reduce_by(plan, pairwise = FALSE, append = FALSE)
-  expect_equivalent(z[, c("target", "command")], y[nrow(y), ])
+  equivalent_plans(z[, c("target", "command")], y[nrow(y), ])
   x <- reduce_by(
     plan, m___from,
     prefix = "xyz",
@@ -330,8 +328,9 @@ test_with_dir("reduce_by()", {
     n__ = as.character(NA),
     n___from = as.character(NA)
   )
+  y <- sanitize_plan(y)
   expected <- bind_plans(plan, y)
-  expect_equivalent(x[order(x$target), ], expected[order(expected$target), ])
+  equivalent_plans(x[order(x$target), ], expected[order(expected$target), ])
   x <- reduce_by(
     plan, m___from,
     prefix = "xyz",
@@ -351,8 +350,9 @@ test_with_dir("reduce_by()", {
     n__ = as.character(NA),
     n___from = as.character(NA)
   )
+  y <- sanitize_plan(y)
   expected <- bind_plans(plan, y)
-  expect_equivalent(x[order(x$target), ], expected[order(expected$target), ])
+  equivalent_plans(x[order(x$target), ], expected[order(expected$target), ])
   x <- reduce_by(
     plan, m___from, prefix = "xyz", op = ", ", begin = "c(", end = ")",
     pairwise = FALSE, append = TRUE
@@ -365,8 +365,9 @@ test_with_dir("reduce_by()", {
     n__ = as.character(NA),
     n___from = as.character(NA)
   )
+  y <- sanitize_plan(y)
   expected <- bind_plans(plan, y)
-  expect_equivalent(x[order(x$target), ], expected[order(expected$target), ])
+  equivalent_plans(x[order(x$target), ], expected[order(expected$target), ])
   x <- reduce_by(plan, m___from, n___from, append = TRUE)
   y <- weak_tibble(
     target = c(
@@ -386,8 +387,9 @@ test_with_dir("reduce_by()", {
     n__ = as.character(NA),
     n___from = c(rep(NA, 3), "y")
   )
+  y <- sanitize_plan(y)
   expected <- bind_plans(plan, y)
-  expect_equivalent(x[order(x$target), ], expected[order(expected$target), ])
+  equivalent_plans(x[order(x$target), ], expected[order(expected$target), ])
   x <- reduce_by(plan, m___from, n___from, pairwise = FALSE, append = TRUE)
   y <- weak_tibble(
     target = c(
@@ -405,8 +407,9 @@ test_with_dir("reduce_by()", {
     n__ = as.character(NA),
     n___from = c(NA, "y", NA)
   )
+  y <- sanitize_plan(y)
   expected <- bind_plans(plan, y)
-  expect_equivalent(x[order(x$target), ], expected[order(expected$target), ])
+  equivalent_plans(x[order(x$target), ], expected[order(expected$target), ])
   plan$from <- c(rep("x", 4), rep("y", 2), NA)
   x <- reduce_by(
     plan,
@@ -416,9 +419,7 @@ test_with_dir("reduce_by()", {
     pairwise = FALSE,
     filter = from == "y"
   )
-  y <- weak_tibble(
-    target = c(plan$target, "xyz_y"),
-    command = c(plan$command, "y_a + y_b")
-  )
-  expect_equivalent(x[, c("target", "command")], y)
+  new_row <- drake_plan(xyz_y = y_a + y_b)
+  y <- bind_plans(plan, new_row)
+  equivalent_plans(x[, c("target", "command")], y)
 })